requirements.txt
django>=5.0.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
pillow
gunicorn
psycopg2-binary
dj-database-url
whitenoise
cloudinary
django-cloudinary-storage
